Don't add jsf to feature list when jsfContainer is present
This change adds jsfContainer to the list of matched features whenever javax.faces.* is detected, but sets jsf as the preferred feature. Doing so causes feature resolution to be satisfied by the existence of jsfContainer-2.2 in the feature list when seeing the javax.faces import, meaning it does not try to add jsf-2.2 to the feature list like it does without this change. It does not affect the behavior of feature resolution when no jsf feature is present.