Issue #29: Add an error message for when a module fails to publish
When a module fails to publish in OLT, only a NPE will be thrown,
without any further useful information being shown. To fix this, I've
surrounded the function that throws the NPE with a try/catch and when an
NPE is caught, a message and a stack trace will be printed out.