Detect jsf and add jsfContainer as alternate feature (#95)
* Detect jsf and add jsfContainer as alternate

* Tweak name matching and refactor into helper method

* Create test class for alternative feature

The test first adds the app in question to an unmodified server and checks that the expected feature is added. It then removes the app, and adds the acceptable alternative feature to the config, then adds the app again. Then the test checks that the alternative feature remains, and the main feature, that would be added automatically otherwise, was not added.

* Add copyright statements and clean up test files